requirements.txt
fastapi
uvicorn
streamlit

pydantic
python-dotenv

pypdf
pdfplumber
pytesseract
Pillow

httpx
tenacity

beautifulsoup4
lxml

loguru
pytest
pytest-cov